Revert "Refactor GPU executable implementation"
This reverts commit cfe75246f4e5937e8d5c4e98e5b5b6113d171e27.
